+++ fractal_server/app/schemas/v2/task_group.py
@@ -5,7 +5,6 @@ from typing import Optional
 from pydantic import BaseModel
 from pydantic import Extra
 from pydantic import Field
-from pydantic import root_validator
 from pydantic import validator
 
 from .._validators import val_absolute_path
@@ -76,21 +75,6 @@ class TaskGroupCreateV2Strict(TaskGroupCreateV2):
     venv_path: str
     version: str
     python_version: str
-
-    @root_validator
-    def check_wheel_file(cls, values):
-        origin = values.get("origin")
-        wheel_path = values.get("wheel_path")
-        bad_condition_1 = (
-            origin == TaskGroupV2OriginEnum.WHEELFILE and wheel_path is None
-        )
-        bad_condition_2 = (
-            origin != TaskGroupV2OriginEnum.WHEELFILE
-            and wheel_path is not None
-        )
-        if bad_condition_1 or bad_condition_2:
-            raise ValueError(f"Cannot have {origin=} and {wheel_path=}.")
-        return values
 
 
 class TaskGroupReadV2(BaseModel):