+++ fractal_server/config.py
@@ -304,8 +304,7 @@ class Settings(BaseSettings):
     """
     Logging-level threshold for logging
 
-    Only logs of with this level (or higher) will appear in the console logs;
-    see details [here](../internals/logs/).
+    Only logs of with this level (or higher) will appear in the console logs.
     """
 
     FRACTAL_LOCAL_CONFIG_FILE: Optional[Path]
@@ -494,6 +493,39 @@ class Settings(BaseSettings):
     """
     Whether to include the v1 API.
     """
+
+    FRACTAL_PIP_CACHE_DIR: Optional[str] = None
+    """
+    Absolute path to the cache directory for `pip`; if unset,
+    `--no-cache-dir` is used.
+    """
+
+    @validator("FRACTAL_PIP_CACHE_DIR", always=True)
+    def absolute_FRACTAL_PIP_CACHE_DIR(cls, v):
+        """
+        If `FRACTAL_PIP_CACHE_DIR` is a relative path, fail.
+        """
+        if v is None:
+            return None
+        elif not Path(v).is_absolute():
+            raise FractalConfigurationError(
+                f"Non-absolute value for FRACTAL_PIP_CACHE_DIR={v}"
+            )
+        else:
+            return v
+
+    @property
+    def PIP_CACHE_DIR_ARG(self) -> str:
+        """
+        Option for `pip install`, based on `FRACTAL_PIP_CACHE_DIR` value.
+
+        If `FRACTAL_PIP_CACHE_DIR` is set, then return
+        `--cache-dir /somewhere`; else return `--no-cache-dir`.
+        """
+        if self.FRACTAL_PIP_CACHE_DIR is not None:
+            return f"--cache-dir {self.FRACTAL_PIP_CACHE_DIR}"
+        else:
+            return "--no-cache-dir"
 
     FRACTAL_MAX_PIP_VERSION: str = "24.0"
     """