+++ fractal_server/tasks/v2/utils_templates.py
@@ -19,9 +19,9 @@ def customize_template(
     Customize a bash-script template and write it to disk.
 
     Args:
-        template_filename:
-        templates_folder:
-        replacements:
+        template_name: Name of the template that will be customized.
+        replacements: List of replacements for template customization.
+        script_path: Local path where the customized template will be written.
     """
     # Read template
     template_path = TEMPLATES_DIR / template_name
@@ -81,6 +81,7 @@ def get_collection_replacements(
             "__FRACTAL_MAX_PIP_VERSION__",
             settings.FRACTAL_MAX_PIP_VERSION,
         ),
+        ("__FRACTAL_PIP_CACHE_DIR_ARG__", settings.PIP_CACHE_DIR_ARG),
         (
             "__PINNED_PACKAGE_LIST__",
             task_group.pinned_package_versions_string,