+++ tests/v2/01_schemas/test_task_collection.py
@@ -31,11 +31,6 @@ def test_TaskCollectPipV2():
     )
     assert sanitized_keys.pinned_package_versions == dict(a="1.0.0")
 
-    with pytest.raises(
-        ValidationError, match="Local-package path must be absolute"
-    ):
-        TaskCollectPipV2(package="not/absolute.whl")
-
     with pytest.raises(ValidationError):
         TaskCollectPipV2(
             package="pkg", pinned_package_versions={";maliciouscmd": "1.0.0"}
@@ -69,7 +64,6 @@ def test_TaskCollectPipV2():
 
 
 async def test_TaskCollectCustomV2(testdata_path):
-
     manifest_file = (
         testdata_path.parent
         / "v2/fractal_tasks_mock"
@@ -201,30 +195,6 @@ def test_TaskGroupCreateV2Strict():
             version="c",
             origin=TaskGroupV2OriginEnum.WHEELFILE,
             wheel_path="/a",
-            pkg_name="x",
-            user_id=1,
-        )
-    # Wheel path set for pypi origin
-    with pytest.raises(ValueError, match="origin"):
-        TaskGroupCreateV2Strict(
-            path="/a",
-            venv_path="/b",
-            version="c",
-            python_version="d",
-            origin=TaskGroupV2OriginEnum.PYPI,
-            wheel_path="/a",
-            pkg_name="x",
-            user_id=1,
-        )
-    # Wheel path unset for wheel origin
-    with pytest.raises(ValueError, match="origin"):
-        TaskGroupCreateV2Strict(
-            path="/a",
-            venv_path="/b",
-            version="c",
-            python_version="d",
-            origin=TaskGroupV2OriginEnum.WHEELFILE,
-            wheel_path=None,
             pkg_name="x",
             user_id=1,
         )